Made a new method in ast class, modified a method in lexer and encountered a bug when trying to debug
Made the getRoot() method in the ast class, intended to return the root
of a tree you created, also modified the return variable in the medthod
readLine() in lexer class. Started to make a file for testing puposes,
which I excluded from getting to the git, but encountered a bug when
trying to use it. Tomorrow I will fix it, hopefully